added output path to lca
