+++ scripts/06-blast-16S-MiFish.py
@@ -1,178 +1,129 @@
 #!/usr/bin/env python3
 import argparse
-import subprocess
 import pandas as pd
-import pytaxonkit
+import numpy as np
+import pytaxonkit as ptk
 
 '''
-Script to run blast with custom 16S and MiFish databases
+Lowest common ancestor script for blast results of 16S and MiFish
 
-This script will run blastn on the custom databases and return taxonomic annotations.
-The script requires pandas and pytaxonkit in either an activated conda environment or installed on the system via
-e.g. "pip install pandas" and "conda install -c bioconda pytaxonkit".
-
-This pipeline contains the conda environmnent required for this script, which needs to be activated before running it.
-
-It further requires a conda environment with blast installed, named 'blast-2.12.0'
-
-Lastly, the taxonkit taxdump file will need to be installed:
-mkdir ~/.taxonkit
-cd ~/.taxonkit
-wget ftp://ftp.ncbi.nih.gov/pub/taxonomy/taxdump.tar.gz
-tar xzvf taxdump.tar.gz
+Filtering done by maximum percent identity and maximum length
 
 INPUT:
-      - fasta file with ASV sequences from DADA2, containing the ASV IDs and sequences
-      - output file name/path
-      - database name either 16S or MiFish
-
+     Results from the script 06-blast-16S-MiFish.py (blast results)
 
 RETURNS:
-*******************************************
-Blast against custom 16S/MiFish database finished!
-Output file : Output_file_name.tsv
-Glimpse at it:
-*******************************************
-
-      ASV                  Species   taxid  pident  length  mismatch  gaps  qstart  qend  sstart  send        evalue  qcovs  qcovhsp
-0  ASV_64       Parupeneus indicus  334904  88.298     188        20     1       1   186     201    14  2.920000e-61    100      100
-1  ASV_64       Parupeneus indicus  334904  88.298     188        20     1       1   186     201    14  2.920000e-61    100      100
-2  ASV_64  Parupeneus heptacanthus  334900  89.941     169        17     0       1   169     200    32  1.360000e-59     91       91
-3  ASV_64      Priacanthus tayenus  443711  78.659     164        30     5       1   161     202    41  3.970000e-25     87       87
-4  ASV_64      Priacanthus tayenus  443711  78.659     164        30     5       1   161     202    41  3.970000e-25     87       87
-
+    Lowest common ancestor table based on percent identity and sequencing length, combined with taxonit lca call
 
 Usage:
-     python 06-blast-16S-MiFish.py --dada2_file [DADA2 fasta file] \
-                                   --out_path [filename or path for output] \
-                                   --database [either '16S' or 'MiFish']
+    python lca_16S_MiFish.py --blast_results [path to output from 16S/MiFish blast results] \
+                             --dada2_asv_table [path to dada2 ASV count table output] \
+                             --output [path to output file]
+
+
 '''
-
-#............................................................................
-# Code to activate conda environments
-CONDA_BLAST    = 'eval "$(conda shell.bash hook)" ; conda activate blast-2.12.0'
-
-# Locations of 16S and MiFish databases
-DB16S  = "/DATA/sandbox/amplicon_databases/16SDB/Bunce_16S.fasta"
-MIFISH = "/DATA/sandbox/amplicon_databases/MiFishDB/MiFishdb.fasta"
-#............................................................................
 
 def main():
 
     OUT_PATH="./"
-    parser = argparse.ArgumentParser(description='Run blastn on a query ASV sequence from a DADA2 run and return taxonomic annotation.')
-    parser.add_argument('--dada2_file', metavar='FILE', type=str,
-                        help='''Filename of dada2 fasta file with query sequences of ASVs''')
-    parser.add_argument('--out_path', metavar='DIR', type=str, default=OUT_PATH,
-                        help=f'Output filename, defaulting to "{OUT_PATH}".')
-    parser.add_argument('--database', metavar='DB', type=str, choices=['16S', 'MiFish'],
-                        help='''Select either the 16S or the MiFish database.''')
+    parser = argparse.ArgumentParser(description='Lowest Common Ancestor retrieval for blast results')
+    parser.add_argument('--blast_results', metavar="BLAST_RES", type=str,
+                        help='''Filename and path of blast output'''),
+    parser.add_argument('--dada2_asv_table', metavar="DADA2_RES", type=str,
+                       help='''Filename and path to dada2 ASV count table''')
+    parser.add_argument('--output', metavar="OUTPUT", type=str, default=OUT_PATH,
+                        help=f'Path to output file defaulting to "{OUT_PATH}".')
+
     args = parser.parse_args()
-    dada2_file = args.dada2_file
-    out_path = args.out_path
-    database = args.database
+    blast_results = args.blast_results
+    dada2_asv = args.dada2_asv_table
+    output = args.output + "_LCA.csv"
 
-    OUT_NAME = out_path + database + '_blast_results.tsv'
+    # Read the blast output and the dada2 ASV table
+    dat = pd.read_csv(blast_results, sep = '\t')
+    dada2 = pd.read_csv(dada2_asv)
 
+    filtered_data = filter_pident_length(dat)
+    assign_lineage(filtered_data)
 
-    # Blast either against the 16S or MiFish database and return formatted file
-    if database == '16S':
-        # Run blastn for the 16S database
-        blastn(dada2_file, OUT_NAME, database)
-        # Process the result file to retrieve the correct format
-        process_blast_output(OUT_NAME, database)
-    else:
-        # Run blastn for the MiFish database
-        blastn(dada2_file, OUT_NAME, database)
-        # Process the result file to retrieve the correct format
-        process_blast_output(OUT_NAME, database)
+    # Merge the LCA and the DADA2 results
+    print(pd.merge(lca, dada2.set_index('sample_id').transpose().rename_axis('ASV').reset_index(), on='ASV', how='right').head())
+    final_lca = pd.merge(lca, dada2.set_index('sample_id').transpose().rename_axis('ASV').reset_index(), on='ASV', how='right')
+
+    final_lca.to_csv(output, index = False)
 
 
 
-# Define function to run blast on 16S and MiFish database
-
-def blastn(dada2_file, out_file, database):
-
+def filter_pident_length(data):
     '''
-    blastn function
+    Filter for highest percent identity of query and database sequence as well as sequence length
 
-    Function to activate the blast conda environment and blast a query sequence
-    against a custom 16S and the MiFish database http://mitofish.aori.u-tokyo.ac.jp/download.html
+    For the LCA, we need to filter the blast results; this is a rough default filter that only retains
+    the entries per ASV that have the highest pident and length entries
 
     Parameters:
-    dada2_fasta (path): query amplicon sequences from the DADA2 R script
-    out_file (string) : name and path for output file
-    '''
-    if database == '16S':
-        blast_command = 'blastn -db ' + DB16S + ' -query ' + dada2_file + ' -outfmt "6 stitle qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue itscore qcovs qcovhsp" -num_threads 200 > ' + out_file
-        subprocess.call(CONDA_BLAST + ';' + blast_command, shell = True)
-
-    else:
-        blast_command = 'blastn -db ' + MIFISH + ' -query ' + dada2_file + ' -outfmt "6 stitle qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue itscore qcovs qcovhsp" -num_threads 200 > ' + out_file
-        subprocess.call(CONDA_BLAST + ';' + blast_command, shell = True)
-
-
-
-# Prepare the output from the 16S database blast for taxonkit
-
-def process_blast_output(out_file, database):
+    data: data set loaded from the path provided to the main function
 
     '''
-    Internal function to prepare the 16S query file for taxonkit annotation
+    print('**************************')
+    print('Filtering blast results...')
+    print('**************************')
 
-    Parameters:
-    out_file (string): taken from the command line, the blastn output file name taken as input for this function
+    lca = pd.DataFrame()
+    for i in data.ASV.unique():
+
+        dat_slice = data[data.ASV == i].copy()
+        dat_slice['filter'] = dat_slice['pident'] + dat_slice['length']
+        output = dat_slice[dat_slice['filter'] == dat_slice['filter'].max()]
+        lca = pd.concat([lca, output])
+
+
+    return lca.reset_index(drop=True)
+
+
+def assign_lineage(filtered_data):
+
     '''
 
-    if database == '16S':
+    '''
 
-        blast_16 = pd.read_csv(out_file, sep = '\t', header = None)
-        blast_16[blast_16.columns[0]] = blast_16[blast_16.columns[0]].str.split(',').str[0].values
-        blast_16['taxid'] = pytaxonkit.name2taxid(blast_16[blast_16.columns[0]])['TaxID']
-        blast_16 = blast_16[[1, 0, 'taxid', 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]]
-        blast_16.columns = ['ASV', 'Species', 'taxid', 'pident',  'length', 'mismatch', 'gaps', 'qstart', 'qend', 'sstart', 'send', 'evalue','qcovs', 'qcovhsp' ]
+    print('**************************')
+    print('Taxonkit lca algorithm  ..')
+    print('**************************')
 
-        blast_16.to_csv(out_file, sep = '\t', index = False)
-        print('')
-        print('*******************************************')
-        print('Blast against custom 16S database finished!')
-        print('Output file : ' + out_file)
-        print('Glimpse at it:')
-        print('*******************************************')
-        print('')
+    filtered_data['taxid'] = filtered_data['taxid'].fillna(-1)
+    filtered_data['taxid'] = filtered_data['taxid'].astype(int)
+    filtered_data['taxid'] = filtered_data['taxid'].astype(str)
+    filtered_data['taxid'] = filtered_data['taxid'].replace('-1', np.nan)
 
-        print(blast_16.head())
+    #lineage = ptk.lineage(filtered_data.taxid.tolist())
+    #filtered_data['FullLineage'] = lineage[['FullLineage']]
 
-    else:
+    lca = pd.DataFrame()
 
-        taxa_list = []
-        blast_mifish = pd.read_csv(out_file, sep = '\t', header = None)
+    for i in filtered_data.ASV.unique():
+        dat_slice = filtered_data[filtered_data.ASV ==i].copy()
+        # Because some entries in the 16S and MiFish databases do not map to a taxid,
+        # this results in some NaN entries. This if else statement will handle those and
+        # skip it by adding NaN to the output
+        if dat_slice.taxid.isna().all() == True:
+            out = pd.DataFrame(dat_slice.taxid)
+        else:
+            out = pd.DataFrame([ptk.lca(dat_slice.taxid.tolist())])
+            out = out.assign(ASV=i)
 
-        for hit in blast_mifish[blast_mifish.columns[0]]:
+            lca = pd.concat([lca, out])
 
-            if hit.startswith('gi'):
-                tax = hit.split('|')[6].strip()
-                taxa_list.append(tax)
+    # Need to reset the index, otherwise the additon of the lineage column does not work, as all indices in lca ar '0'
+    lca = lca.reset_index(drop=True)
+    lineage = ptk.lineage(lca[0].values.tolist())
+    lca['LCA'] = lineage['Name']
+    lca = lca.rename(columns = {0 : 'taxid'})
+    #lca.colunms = ['taxid', 'ASV', 'LCA']
+    lca = lca[['ASV', 'taxid', 'LCA']]
+    #lineage = ptk.lineage(lca[[]])
+    #lca.to_csv('lca_results_test.tsv', sep='\t', index = False)
+    return lca
 
-            else:
 
-                tax = hit.split('|')[2].split('(')[0].strip()
-                taxa_list.append(tax)
-
-        blast_mifish['taxa'] = taxa_list
-        blast_mifish['taxid'] = pytaxonkit.name2taxid(blast_mifish[blast_mifish.columns[len(blast_mifish.columns)-1]])['TaxID']
-
-        blast_mifish = blast_mifish[[1, 'taxa', 'taxid', 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]]
-        blast_mifish.columns = ['ASV', 'Species', 'taxid', 'pident',  'length', 'mismatch', 'gaps', 'qstart', 'qend', 'sstart', 'send', 'evalue','qcovs', 'qcovhsp' ]
-        blast_mifish.to_csv(out_file, sep = '\t', index = False)
-
-        print('')
-        print('*******************************************')
-        print('Blast against custom MiFish database finished!')
-        print('Output file : ' + out_file)
-        print('Glimpse at it:')
-        print('*******************************************')
-        print('')
-
-        print(blast_mifish.head())
 main()
